LC-plus-one [Runtime: N/A; Memory: N/A]

+++ array/plus-one.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    vector<int> getArray(int sum){
+    vector<int> getArray(long long sum){
         vector<int>ans;
         while(sum != 0){
             ans.push_back(sum%10);
@@ -10,7 +10,7 @@ public:
         return ans;
     }
     vector<int> plusOne(vector<int>& digits) {
-        int sum = 0;
+        long long sum = 0;
         for(auto i: digits) sum= sum *10 + i;
         sum+=1;
         return getArray(sum);
